Change all BigInteger usages to Longs

+++ src/main/kotlin/Day06.kt
@@ -1,5 +1,3 @@
-import java.math.BigInteger
-
 fun main() {
     val input = getInputLines("day6")
 
@@ -7,9 +5,9 @@ fun main() {
     println(part2(input))
 }
 
-private fun part1(input: List<String>): BigInteger {
+private fun part1(input: List<String>): Long {
     val operations = input.last().trim().split(Regex("\\s+"))
-    val values = input.dropLast(1).map { line -> line.trim().split(Regex("\\s+")).map { it.toBigInteger() } }
+    val values = input.dropLast(1).map { line -> line.trim().split(Regex("\\s+")).map { it.toLong() } }
     return operations.indices.sumOf {
         if (operations[it] == "+") {
             values.sumOf { value -> value[it] }
@@ -19,14 +17,14 @@ private fun part1(input: List<String>): BigInteger {
     }
 }
 
-private fun part2(input: List<String>): BigInteger {
+private fun part2(input: List<String>): Long {
     val operations = input.last().split(Regex("\\s(?=[*+])"))
     var remainingValues = input.dropLast(1)
     return operations.indices.sumOf { i ->
         val currentOperation = operations[i]
         val currentColumns = remainingValues.map { it.take(currentOperation.length) }
         val currentValues = currentOperation.indices
-            .map { j -> currentColumns.map { it[j] }.joinToString("").trim().toBigInteger() }
+            .map { j -> currentColumns.map { it[j] }.joinToString("").trim().toLong() }
         remainingValues = remainingValues.map { it.drop(currentOperation.length + 1) }
         if (currentOperation.startsWith("+")) {
             currentValues.sumOf {it}
